Making links optional

+++ deploy.sh
@@ -1,14 +1,27 @@
 #!/bin/bash
 
+# Whether to deploy the frontend
 frontend=false
+# What release of the frontend to deploy
 frontendRelease=LATEST
+# The Git location for the frontend
 frontendGit="git:osmlab/maproulette3"
+# Whether to deploy the API
 api=false
+# What release of the API to deploy
 apiRelease=LATEST
+# The Git location for the API
 apiGit="git:maproulette/maproulette2"
+# Whether to wipe the docker database, start clean
 wipeDB=false
+# What port to expose the docker database on, by default will not expose it
 dbPort=""
+# What host the API is on, used for Swagger
 apiHost="maproulette.org"
+# Whether the database being used is external or not. If it is external than won't link and build the database images
+dbExternal=false
+# Whether to link the API and frontend, would only not use this if connecting to another API
+apiExternal=false
 
 while true; do
     case "$1" in
@@ -61,6 +74,12 @@ while true; do
             apiHost=$2
             shift
         ;;
+        --dbExternal)
+            dbExternal=true
+        ;;
+        --apiExternal)
+            apiExternal=true
+        ;;
         *)
             break
         ;;
@@ -71,34 +90,36 @@ done
 echo "API: $api $apiRelease $apiGit"
 echo "FRONTEND: $frontend $frontendRelease $frontendGit"
 if [[ "$api" = true ]]; then
-    echo "deploying database..."
-    if [[ "$wipeDB" == true ]]; then
-        echo "Stopping and removing mr-postgis container"
-        docker stop mr-postgis
-        docker rm mr-postgis
-    fi
+    if [[ "$dbExternal" != true ]]; then
+        echo "deploying database..."
+        if [[ "$wipeDB" == true ]]; then
+            echo "Stopping and removing mr-postgis container"
+            docker stop mr-postgis
+            docker rm mr-postgis
+        fi
 
-    instance=$(docker ps -a | grep mdillon/postgis)
-    if [[ -z "$instance" ]]; then
-        echo "Building new mr-postgis container"
-        docker run $dbPort \
-            --name mr-postgis \
-            -e POSTGRES_DB=mrdata \
-            -e POSTGRES_USER=mrdbuser \
-            -e POSTGRES_PASSWORD=mrdbpass \
-            -d mdillon/postgis
-        sleep 10
-    fi
-    instance=$(docker ps | grep mdillon/postgis)
-    if [[ -z "$instance" ]]; then
-        echo "Restarting mr-postgis container"
-        docker start mr-postgis
+        instance=$(docker ps -a | grep mdillon/postgis)
+        if [[ -z "$instance" ]]; then
+            echo "Building new mr-postgis container"
+            docker run $dbPort \
+                --name mr-postgis \
+                -e POSTGRES_DB=mrdata \
+                -e POSTGRES_USER=mrdbuser \
+                -e POSTGRES_PASSWORD=mrdbpass \
+                -d mdillon/postgis
+            sleep 10
+        fi
+        instance=$(docker ps | grep mdillon/postgis)
+        if [[ -z "$instance" ]]; then
+            echo "Restarting mr-postgis container"
+            docker start mr-postgis
+        fi
     fi
     echo "deploying api..."
-    ./backend/docker.sh $apiRelease $apiGit $apiHost
+    ./api/docker.sh $apiRelease $apiGit $apiHost $dbExternal
     sleep 10
 fi
 if [[ "$frontend" = true ]]; then
     echo "deploying frontend..."
-    ./frontend/docker.sh $frontendRelease $frontendGit
+    ./frontend/docker.sh $frontendRelease $frontendGit $apiExternal
 fi
